requirements.txt
nextcord==2.3.2
pymongo==4.3.2
python-dotenv==0.21.0
matplotlib==3.6.2